Reorganize metabackend/model/ files.
